+++ oid4vp-java/oid4vp-core/src/main/java/de/eecc/oid4vc/oid4vp/exception/InternalError.java
@@ -1,0 +1,14 @@
+package de.eecc.oid4vc.oid4vp.exception;
+
+public record InternalError(String detail) implements Oid4VpError {
+
+    @Override
+    public String message() {
+        return detail;
+    }
+
+    @Override
+    public int suggestedHttpStatus() {
+        return 500;
+    }
+}