+++ oid4vp-java/oid4vp-core/src/main/java/de/eecc/oid4vc/oid4vp/exception/UnknownState.java
@@ -1,0 +1,14 @@
+package de.eecc.oid4vc.oid4vp.exception;
+
+public record UnknownState() implements Oid4VpError {
+
+    @Override
+    public String message() {
+        return "Unknown or expired state";
+    }
+
+    @Override
+    public int suggestedHttpStatus() {
+        return 400;
+    }
+}